// Frontend/workload-management/src/app/services/models/academic-rank-details-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { AcademicRank } from '../models/academic-rank';
import { Semester } from '../models/semester';
export interface AcademicRankDetailsRequest {
  academicRank: AcademicRank;
  contactHoursForFullTime: number;
  cpForFullTime: number;
  salary: number;
  semester: Semester;
}